requirements.txt
aiohttp
PyYAML
customtkinter
colorama
